Add initial host / join game support
Hosting a game adds a game to a list of games that are being hosted.
Joining a game gets back a list of the current games that are being
hosted, and adds the user to a list of users that will be notified every
time a new game is hosted.